Extract live send run asset planning
