+++ src/PlateauResoniteLink/Targets/Resonite/ResoniteQueuedCityObjectSender.cs
@@ -1,0 +1,440 @@
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Globalization;
+using System.IO;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+
+using PlateauResoniteLink.Application.Importing;
+using PlateauResoniteLink.Application.Logging;
+using PlateauResoniteLink.Domain.Importing;
+using PlateauResoniteLink.Transport.ResoniteLink;
+
+namespace PlateauResoniteLink.Targets.Resonite;
+
+internal interface IResoniteQueuedCityObjectSender
+{
+    Task SendAsync(
+        LiveSendRunState state,
+        IResoniteLinkClient routedClient,
+        LiveSendQueuedCityObject queuedCityObject,
+        ResoniteLinkSendDiagnostics diagnostics,
+        Action<string>? progressReporter,
+        CancellationToken cancellationToken);
+}
+
+internal sealed class ResoniteQueuedCityObjectSender(
+    ITerrainTextureAssetGenerator terrainTextureAssetGenerator,
+    Execution.IResoniteDatasetLicenseWriter datasetLicenseWriter,
+    IResonitePreparedCityObjectImporter preparedCityObjectImporter) : IResoniteQueuedCityObjectSender
+{
+    private readonly ITerrainTextureAssetGenerator terrainTextureAssetGenerator =
+        terrainTextureAssetGenerator ?? throw new ArgumentNullException(nameof(terrainTextureAssetGenerator));
+    private readonly Execution.IResoniteDatasetLicenseWriter datasetLicenseWriter =
+        datasetLicenseWriter ?? throw new ArgumentNullException(nameof(datasetLicenseWriter));
+    private readonly IResonitePreparedCityObjectImporter preparedCityObjectImporter =
+        preparedCityObjectImporter ?? throw new ArgumentNullException(nameof(preparedCityObjectImporter));
+
+    [System.Diagnostics.CodeAnalysis.SuppressMessage(
+        "Design",
+        "CA1031:Do not catch general exception types",
+        Justification = "Live send should log and skip individual city object send failures while keeping the lane alive.")]
+    public async Task SendAsync(
+        LiveSendRunState state,
+        IResoniteLinkClient routedClient,
+        LiveSendQueuedCityObject queuedCityObject,
+        ResoniteLinkSendDiagnostics diagnostics,
+        Action<string>? progressReporter,
+        CancellationToken cancellationToken)
+    {
+        ArgumentNullException.ThrowIfNull(state);
+        ArgumentNullException.ThrowIfNull(routedClient);
+        ArgumentNullException.ThrowIfNull(queuedCityObject);
+        ArgumentNullException.ThrowIfNull(diagnostics);
+
+        Interlocked.Increment(ref state.Progress.AttemptedCityObjectCount);
+        try
+        {
+            PreparedCityObject preparedCityObject = await AwaitWithSlowCityObjectWarningAsync(
+                CreatePreparationTask(
+                    state,
+                    routedClient,
+                    queuedCityObject.CityObject,
+                    diagnostics,
+                    progressReporter,
+                    cancellationToken),
+                cancellationToken);
+            await preparedCityObjectImporter.ImportAsync(
+                state,
+                routedClient,
+                queuedCityObject,
+                preparedCityObject,
+                diagnostics,
+                progressReporter,
+                cancellationToken);
+
+            int processedCount = Interlocked.Increment(ref state.Progress.ProcessedCityObjectCount);
+            ReportProgress(
+                progressReporter,
+                PlateauLog.Info(
+                    "live",
+                    $"Sent city object {processedCount}: "
+                    + $"{preparedCityObject.CityObject.DisplayName} "
+                    + $"({preparedCityObject.CityObject.PackageName}/{preparedCityObject.CityObject.SlotKey})"));
+        }
+        catch (OperationCanceledException)
+        {
+            throw;
+        }
+        catch (Exception exception)
+        {
+            if (!IsRecoverableCityObjectSendFailure(exception))
+            {
+                throw;
+            }
+
+            int failedCount = Interlocked.Increment(ref state.Progress.FailedCityObjectCount);
+            ReportProgress(
+                progressReporter,
+                PlateauLog.Warning(
+                    "live",
+                    $"Skipping city object after send failure {failedCount}: "
+                    + $"{queuedCityObject.CityObject.DisplayName} "
+                    + $"({queuedCityObject.CityObject.PackageName}/{queuedCityObject.CityObject.SlotKey}). "
+                    + $"Reason: {exception.Message}"));
+        }
+        finally
+        {
+            await queuedCityObject.MemoryLease.DisposeAsync();
+        }
+    }
+
+    private Task<PreparedCityObject> CreatePreparationTask(
+        LiveSendRunState state,
+        IResoniteLinkClient routedClient,
+        ResoniteConstructionCityObject cityObject,
+        ResoniteLinkSendDiagnostics diagnostics,
+        Action<string>? progressReporter,
+        CancellationToken callerCancellationToken)
+    {
+        if (Interlocked.CompareExchange(ref state.Progress.FirstCityObjectPreparationStartedLogged, 1, 0) == 0)
+        {
+            ReportProgress(
+                progressReporter,
+                PlateauLog.Info(
+                    "live",
+                    $"City object preparation started after {state.Runtime.ElapsedTotalSeconds:F3}s: "
+                    + $"{cityObject.DisplayName} ({cityObject.PackageName}/{cityObject.SlotKey}) "
+                    + $"mesh='{cityObject.ActualMeshCode}'."));
+        }
+
+        CancellationToken processingCancellationToken = state.Runtime.ProcessingCancellationToken;
+        return PrepareCityObjectWithLinkedCancellationAsync(
+            state,
+            routedClient,
+            cityObject,
+            diagnostics,
+            progressReporter,
+            callerCancellationToken,
+            processingCancellationToken);
+    }
+
+    private async Task<PreparedCityObject> PrepareCityObjectWithLinkedCancellationAsync(
+        LiveSendRunState state,
+        IResoniteLinkClient routedClient,
+        ResoniteConstructionCityObject cityObject,
+        ResoniteLinkSendDiagnostics diagnostics,
+        Action<string>? progressReporter,
+        CancellationToken callerCancellationToken,
+        CancellationToken processingCancellationToken)
+    {
+        using CancellationTokenSource linkedCancellation = CancellationTokenSource.CreateLinkedTokenSource(
+            callerCancellationToken,
+            processingCancellationToken);
+        return await PrepareCityObjectAsync(
+            state,
+            routedClient,
+            cityObject,
+            diagnostics,
+            progressReporter,
+            linkedCancellation.Token);
+    }
+
+    private async Task<PreparedCityObject> PrepareCityObjectAsync(
+        LiveSendRunState state,
+        IResoniteLinkClient routedClient,
+        ResoniteConstructionCityObject cityObject,
+        ResoniteLinkSendDiagnostics diagnostics,
+        Action<string>? progressReporter,
+        CancellationToken cancellationToken)
+    {
+        cityObject = ResoniteDynamicMaterialUvNormalizer.Normalize(cityObject);
+
+        if (cityObject.Geometry is ResoniteTriangleMeshGeometry triangleGeometry)
+        {
+            ResoniteCityObjectPreparation.ValidateTriangleMeshBindingsForImport(cityObject, triangleGeometry.Mesh);
+        }
+        else if (cityObject.Geometry is ResoniteDynamicTerrainGeometry dynamicTerrain)
+        {
+            ResoniteCityObjectPreparation.ValidateTriangleMeshBindingsForImport(cityObject, dynamicTerrain.StaticMesh.Mesh);
+        }
+
+        (string TerrainMeshCode, TerrainTextureOverlay TerrainOverlay)[] distinctTerrainOverlays = cityObject.Materials
+            .Select((material, materialIndex) => (Material: material, MaterialIndex: materialIndex))
+            .Where(static entry => entry.Material.TerrainOverlay is not null && entry.Material.TerrainMeshCode is not null)
+            .Select(entry => (
+                TerrainMeshCode: ResoniteTerrainOverlayMaterialContract.ValidateMeshCode(
+                    cityObject,
+                    entry.MaterialIndex,
+                    entry.Material,
+                    entry.Material.TerrainMeshCode!,
+                    entry.Material.TerrainOverlay!),
+                TerrainOverlay: entry.Material.TerrainOverlay!))
+            .Distinct()
+            .OrderBy(static entry => entry.TerrainMeshCode, StringComparer.Ordinal)
+            .ThenBy(static entry => entry.TerrainOverlay.PackageName, StringComparer.Ordinal)
+            .ThenBy(static entry => entry.TerrainOverlay.GeographicBounds.MinLatitude)
+            .ThenBy(static entry => entry.TerrainOverlay.GeographicBounds.MinLongitude)
+            .ToArray();
+
+        Task<PreparedTextureReference?>[] terrainOverlayTexturePreparationTasks = distinctTerrainOverlays
+            .Select(entry => PrepareTerrainOverlayTextureReferenceAsync(
+                state,
+                routedClient,
+                progressReporter,
+                entry.TerrainMeshCode,
+                entry.TerrainOverlay,
+                cancellationToken))
+            .ToArray();
+
+        Task<PreparedConstructionGeometry> geometryPreparationTask = cityObject.Geometry switch
+        {
+            ResoniteTriangleMeshGeometry triangleMesh => Task.Run<PreparedConstructionGeometry>(
+                () => ResoniteCityObjectPreparation.PrepareTriangleMeshGeometry(cityObject, triangleMesh.Mesh),
+                cancellationToken),
+            ResoniteTerrainGridGeometry heightMap => Task.Run<PreparedConstructionGeometry>(
+                () => new PreparedTerrainGridGeometry(heightMap, ResoniteCityObjectPreparation.PrepareTerrainGridDisplacementTexture(heightMap)),
+                cancellationToken),
+            ResoniteDynamicTerrainGeometry dynamicTerrain => Task.Run<PreparedConstructionGeometry>(
+                () => new PreparedDynamicTerrainGeometry(
+                    ResoniteCityObjectPreparation.PrepareTriangleMeshGeometry(cityObject, dynamicTerrain.StaticMesh.Mesh),
+                    new PreparedTerrainGridGeometry(
+                        dynamicTerrain.GridMesh,
+                        ResoniteCityObjectPreparation.PrepareTerrainGridDisplacementTexture(dynamicTerrain.GridMesh))),
+                cancellationToken),
+            _ => throw new InvalidOperationException($"Unsupported geometry type '{cityObject.Geometry.GetType().Name}'."),
+        };
+        Stopwatch stopwatch = Stopwatch.StartNew();
+        PreparedTextureReference?[] preparedTextureResults = await Task.WhenAll(
+            terrainOverlayTexturePreparationTasks
+                .Concat(cityObject.Materials
+                    .Where(static material => material.TexturePayload is not null)
+                    .Select(PrepareDirectMaterialTextureReferenceAsync)
+                    .ToArray()));
+        PreparedTextureReference[] preparedTextures = preparedTextureResults
+            .OfType<PreparedTextureReference>()
+            .ToArray();
+        PreparedConstructionGeometry preparedGeometry = await geometryPreparationTask;
+        Dictionary<TerrainTextureOverlay, GeneratedTerrainTexture> preparedTerrainTextureDataByOverlay = preparedTextures
+            .Where(static texture => texture is { TerrainOverlay: not null, GeneratedTerrainTexture: not null })
+            .ToDictionary(
+                static texture => texture.TerrainOverlay!,
+                static texture => texture.GeneratedTerrainTexture!);
+        cityObject = ResoniteCityObjectPreparation.ApplyTerrainTextureCanvasUv(
+            cityObject,
+            preparedTerrainTextureDataByOverlay,
+            clampCanvasUv: ResonitePackageSemantics.IsDemPackage(cityObject.PackageName));
+        if (cityObject.Geometry is ResoniteTriangleMeshGeometry resolvedTriangleMesh
+            && preparedGeometry is PreparedTriangleMeshGeometry)
+        {
+            preparedGeometry = ResoniteCityObjectPreparation.PrepareTriangleMeshGeometry(cityObject, resolvedTriangleMesh.Mesh);
+        }
+        else if (cityObject.Geometry is ResoniteDynamicTerrainGeometry resolvedDynamicTerrain
+            && preparedGeometry is PreparedDynamicTerrainGeometry preparedDynamicTerrain)
+        {
+            preparedGeometry = preparedDynamicTerrain with
+            {
+                StaticMesh = ResoniteCityObjectPreparation.PrepareTriangleMeshGeometry(cityObject, resolvedDynamicTerrain.StaticMesh.Mesh),
+            };
+        }
+        stopwatch.Stop();
+        diagnostics.RecordPrepare(cityObject.PackageName, stopwatch.Elapsed.TotalSeconds);
+
+        if (Interlocked.CompareExchange(ref state.Progress.FirstPreparedCityObjectLogged, 1, 0) == 0)
+        {
+            ReportProgress(
+                progressReporter,
+                PlateauLog.Info(
+                    "live",
+                    $"First city object prepared in {stopwatch.Elapsed.TotalSeconds:F3}s "
+                    + $"after scene start {state.Runtime.ElapsedTotalSeconds:F3}s: "
+                    + $"{cityObject.DisplayName} "
+                    + $"(textures={preparedTextures.Length}, geometry={PreparedConstructionGeometryFormatter.Describe(preparedGeometry)})."));
+        }
+
+        return new PreparedCityObject(
+            cityObject,
+            preparedGeometry,
+            preparedTextures);
+    }
+
+    private async Task<PreparedTextureReference?> PrepareTerrainOverlayTextureReferenceAsync(
+        LiveSendRunState state,
+        IResoniteLinkClient routedClient,
+        Action<string>? progressReporter,
+        string terrainMeshCode,
+        TerrainTextureOverlay terrainTextureOverlay,
+        CancellationToken cancellationToken)
+    {
+        GeneratedTerrainTexture terrainTexture = await terrainTextureAssetGenerator.EnsureTextureAsync(
+            terrainTextureOverlay,
+            cancellationToken);
+        TerrainTextureSource[] usedSources = GetTrackedTerrainTextureSources(terrainTexture, terrainTextureOverlay);
+        foreach (TerrainTextureSource usedSource in usedSources)
+        {
+            int useCount = state.DemSourceUseCounts.AddOrUpdate(
+                usedSource.IdentityKey,
+                1,
+                static (_, current) => checked(current + 1));
+            if (useCount == 1)
+            {
+                ReportProgress(
+                    progressReporter,
+                    PlateauLog.Info(
+                        "live",
+                        $"Resolved DEM terrain texture source for package '{terrainTextureOverlay.PackageName}' "
+                        + $"to {DescribeTerrainTextureSource(usedSource)}."));
+            }
+
+            if (IsGsiFallbackSource(usedSource))
+            {
+                await EnsureGsiFallbackLicenseAsync(state, routedClient, cancellationToken);
+            }
+        }
+
+        return new PreparedTextureReference(
+            TexturePayload: null,
+            TextureSourceKind: ResoniteTextureSourceKind.Dataset,
+            TextureImport: terrainTexture.TextureImport,
+            TerrainMeshCode: terrainMeshCode,
+            TerrainOverlay: terrainTextureOverlay,
+            GeneratedTerrainTexture: terrainTexture);
+    }
+
+    private static TerrainTextureSource[] GetTrackedTerrainTextureSources(
+        GeneratedTerrainTexture terrainTexture,
+        TerrainTextureOverlay terrainTextureOverlay)
+    {
+        if (terrainTexture.UsedSources is { Count: > 0 })
+        {
+            return terrainTexture.UsedSources
+                .Distinct()
+                .ToArray();
+        }
+
+        return
+        [
+            terrainTexture.UsedSource ?? terrainTextureOverlay.PrimarySource,
+        ];
+    }
+
+    private async Task EnsureGsiFallbackLicenseAsync(
+        LiveSendRunState state,
+        IResoniteLinkClient routedClient,
+        CancellationToken cancellationToken)
+    {
+        if (Volatile.Read(ref state.GsiFallbackLicenseEnsured) != 0)
+        {
+            return;
+        }
+
+        await state.GsiFallbackLicenseGate.WaitAsync(cancellationToken);
+        try
+        {
+            if (state.GsiFallbackLicenseEnsured != 0)
+            {
+                return;
+            }
+
+            await datasetLicenseWriter.EnsureGsiFallbackLicenseAsync(
+                routedClient,
+                state.Context.DatasetRootSlot,
+                cancellationToken);
+            Volatile.Write(ref state.GsiFallbackLicenseEnsured, 1);
+        }
+        finally
+        {
+            state.GsiFallbackLicenseGate.Release();
+        }
+    }
+
+    private static bool IsRecoverableCityObjectSendFailure(Exception exception)
+    {
+        return exception is ContinuableImportException
+            || FindResoniteLinkOperationException(exception) is { OperationName: "ImportMesh" or "ImportTexture" or "GetSlot" or "GetComponent" };
+    }
+
+    private static Task<T> AwaitWithSlowCityObjectWarningAsync<T>(
+        Task<T> operationTask,
+        CancellationToken cancellationToken)
+    {
+        return operationTask.WaitAsync(cancellationToken);
+    }
+
+    private static ResoniteLinkOperationException? FindResoniteLinkOperationException(Exception exception)
+    {
+        for (Exception? current = exception; current is not null; current = current.InnerException)
+        {
+            if (current is ResoniteLinkOperationException operationException)
+            {
+                return operationException;
+            }
+        }
+
+        return null;
+    }
+
+    private static bool IsGsiFallbackSource(TerrainTextureSource source)
+    {
+        return DemTerrainTextureDefaults.IsGsiFallbackSource(source);
+    }
+
+    private static string DescribeTerrainTextureSource(TerrainTextureSource source)
+    {
+        return source switch
+        {
+            TerrainTextureGeoReferencedRasterSource rasterSource => string.Create(
+                CultureInfo.InvariantCulture,
+                $"GeoTIFF(path='{Path.GetFileName(rasterSource.SourcePath)}', crs='{rasterSource.Metadata?.CoordinateSystemIdentifier ?? "unknown"}')"),
+            TerrainTextureTileSource tileSource when IsGsiFallbackSource(tileSource) => string.Create(
+                CultureInfo.InvariantCulture,
+                $"GSI seamless photo tile(z={tileSource.ZoomLevel})"),
+            TerrainTextureTileSource tileSource => string.Create(
+                CultureInfo.InvariantCulture,
+                $"PLATEAU-Ortho tile(z={tileSource.ZoomLevel})"),
+            _ => source.GetType().Name,
+        };
+    }
+
+    private static Task<PreparedTextureReference?> PrepareDirectMaterialTextureReferenceAsync(
+        ResoniteMaterialBinding material)
+    {
+        ArgumentNullException.ThrowIfNull(material);
+        ArgumentNullException.ThrowIfNull(material.TexturePayload);
+
+        return Task.FromResult<PreparedTextureReference?>(
+            new PreparedTextureReference(
+                TexturePayload: material.TexturePayload,
+                TextureSourceKind: material.TextureSourceKind,
+                TextureImport: ResoniteTextureImportFactory.CreateRawFromPayload(material.TexturePayload),
+                TerrainMeshCode: null,
+                TerrainOverlay: null));
+    }
+
+    private static void ReportProgress(Action<string>? progressReporter, string message)
+    {
+        progressReporter?.Invoke(message);
+    }
+}
